1.update flvmuxer form LibRtmp-Client-for-Android 2.use libRTMP publish flv

+++ app/src/main/jni/rtmp/librtmp/log.c
@@ -57,7 +57,7 @@ static void rtmp_log_default(int level, const char *format, va_list vl)
 
 	if ( !fmsg ) fmsg = stderr;
 
-	if ( level <= (int)RTMP_debuglevel ) {
+	if ( level <= RTMP_debuglevel ) {
 		if (neednl) {
 			putc('\n', fmsg);
 			neednl = 0;
